refactor(react): add decompose function in utils
