+++ packages/botonic-react/src/components/message.jsx
@@ -2,6 +2,7 @@ import React, { useState, useEffect, useContext } from 'react'
 
 import { isBrowser, isNode } from '@botonic/core'
 import { WebchatContext, RequestContext } from '../contexts'
+import { decomposeComponent } from '../utils'
 import { Button } from './button'
 import { Reply } from './reply'
 
@@ -28,36 +29,13 @@ export const Message = props => {
   const textChildren = React.Children.toArray(children).filter(
     e => ![Button, Reply].includes(e.type)
   )
-
   if (isBrowser()) {
     useEffect(() => {
-      console.log('STRING', textChildren)
-      let carousel = []
-      let img_url = null
-      let new_Child = null
-      try {
-        textChildren[0].props.children.map((e, i) => {
-          let c = e.props.children
-          carousel[i] = {
-            img: c[0].props.src,
-            title: c[1].props.children,
-            subtitle: c[2].props.children,
-            button: c[3].props
-          }
-        })
-      } catch (e) {}
-      if (carousel.length) {
-        console.log('CAAAR;', carousel)
-        new_Child = carousel
-      }
-      try {
-        if (textChildren[0].type == 'img') img_url = textChildren[0].props.src
-      } catch (e) {}
-      if (img_url && !carousel.length) new_Child = img_url
+      let decomposedChildren = decomposeComponent(textChildren)
       let message = {
         id: state.id,
         type,
-        data: new_Child ? new_Child : textChildren,
+        data: decomposedChildren ? decomposedChildren : textChildren,
         from,
         delay,
         typing,
